Fix reply_to property omitted in headers
Adds tests and a fix that 'Reply-To' headers are correctly set, when `reply_to` property of the message object is set (using the initializer argument `reply_to`).

Also, when `reply_to` property is set and header "Reply-To" is also added explictly as a header, then the explicit header value is favored over the message property reply_to, adopting the behaviour of message() in django.core.mail.message.EmailMessage.